+++ org.eclipse.egit.github.core/src/org/eclipse/egit/github/core/service/PullRequestService.java
@@ -453,7 +453,7 @@ public class PullRequestService extends GitHubService {
 	 * @return commit comment
 	 * @throws IOException
 	 */
-	public CommitComment getComment(IRepositoryIdProvider repository, int id)
+	public CommitComment getComment(IRepositoryIdProvider repository, long id)
 			throws IOException {
 		String repoId = getId(repository);
 		StringBuilder uri = new StringBuilder(SEGMENT_REPOS);
@@ -541,7 +541,7 @@ public class PullRequestService extends GitHubService {
 	 * @param id
 	 * @throws IOException
 	 */
-	public void deleteComment(IRepositoryIdProvider repository, int id)
+	public void deleteComment(IRepositoryIdProvider repository, long id)
 			throws IOException {
 		String repoId = getId(repository);
 		StringBuilder uri = new StringBuilder(SEGMENT_REPOS);
